Feature/oce 207 bid extension

+++ persistence-mongodb/src/main/java/org/devgateway/ocds/persistence/mongo/Item.java
@@ -83,7 +83,7 @@ public class Item implements Identifiable {
     @ExcelExport
     @JsonProperty("quantity")
     @Merge(MergeStrategy.ocdsVersion)
-    private Integer quantity;
+    private Double quantity;
 
     /**
      * Description of the unit which the good comes in e.g. hours, kilograms.
@@ -196,7 +196,7 @@ public class Item implements Identifiable {
      *     The quantity
      */
     @JsonProperty("quantity")
-    public Integer getQuantity() {
+    public Double getQuantity() {
         return quantity;
     }
 
@@ -207,7 +207,7 @@ public class Item implements Identifiable {
      *     The quantity
      */
     @JsonProperty("quantity")
-    public void setQuantity(final Integer quantity) {
+    public void setQuantity(final Double quantity) {
         this.quantity = quantity;
     }
